# apps/server/cmd/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API documentation for AgroMart backend services.",
        "title": "AgroMart API",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/api",
    "paths": {
...
        "/sales/orders.csv": {
            "get": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "Streams CSV of aggregated sales orders within an optional date range. Requires admin/manager.",
                "produces": [
                    "text/csv"
                ],
                "tags": [
                    "sales"
                ],
                "summary": "Export sales orders CSV",
                "parameters": [
                    {
                        "type": "string",
                        "description": "From date (YYYY-MM-DD)",
                        "name": "from",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "To date (YYYY-MM-DD)",
                        "name": "to",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "CSV file",
                        "schema": {
                            "type": "file"
                        }
                    },
                    "401": {
                        "description": "invalid tenant/auth",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "403": {
                        "description": "insufficient permissions",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "internal error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
...
    },
...
}